requirements.txt
Django>=3.2,<3.3
wagtail>=2.15,<2.16
flake8==4.0.1
requests==2.27.1
django-environ==0.8.1
